+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/zombiezen/graphql-server/compare/v0.3.0...HEAD
 
+### Fixed
+
+-  Ignore `__typename` in `SelectionSet.OnlyUses`.
+
 ## [0.3.0][]
 
 This release adds the `SelectionSet.OnlyUses` method.